newsproducer/pyproject.toml
[tool.poetry]
name = "newsproducer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.11.9"
kafka-python = "^2.0.2"
newsapi-python = "^0.2.7"
weaviate-client = "4.5.4"
six = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
